# openapi.yaml
openapi: 3.0.1
info:
  title: SwagShop API
  version: 1.0.0
  description: >-
    A RESTful API for E-Commerce back-ends. It uses Node.js, Express.js, and
    PostgreSQL.
...
paths:
  /users:
    post:
      summary: Adds a new user
      operationId: userAdd
      requestBody:
        content:
          application/json:
            example: |-
              {
                "email": "alice@example.com",
                "name": "Alice",
                "password": "53CR3T_P@55W0RD"
              }
      responses:
        '201':
          description: User created
          content:
            application/json:
              example: |-
                {
                  "id": 1,
                  "email": "alice@example.com",
                  "name": "Alice"
                }
        '400':
          description: Malformed or missing data
        default:
          description: Default error sample response
      tags:
        - Users
  /users/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
        description: The id of the user
    get:
      summary: Returns a user
      operationId: userRead
      responses:
        '200':
          description: Successfuly found user
          content:
            application/json:
              example: |-
                {
                  "id": 1,
                  "email": "alice@example.com",
                  "name": "Alice",
                  "admin": true
                }
        '404':
          description: User not found
        default:
          description: Default error sample response
      tags:
        - Users
    patch:
      summary: Updates a user
      operationId: userEdit
      requestBody:
        content:
          application/json:
            example: |-
              {
                "email": "alice@example.net",
                "name": "Alice Anderson",
                "password": "53CR3T_P@55W0RD"
              }
      responses:
        '204':
          description: Successfuly modified user
        '400':
          description: Malformed or missing data
        '404':
          description: User not found
        default:
          description: Default error sample response
      tags:
        - Users
    delete:
      summary: Deletes a user
      operationId: userDelete
      responses:
        '204':
          description: Successfuly deleted user
        '404':
          description: User not found
        default:
          description: Default error sample response
      tags:
        - Users
...
  /users/{id}/orders:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
        description: The id of the user
    get:
      summary: Returns the user's orders
      description: Returns a list containing all of the user's orders
      operationId: usersOrders
      responses:
        '200':
          description: Successfuly retrieved the user's orders
          content:
            application/json:
              example: |-
                [
                  {
                    "id": 1,
                    "datetime": "2024-03-1T14:25:43.511Z",
                    "status": "fulfilled"
                  },
                  {
                    "id": 2,
                    "datetime": "2024-03-2T15:25:43.511Z",
                    "status": "canceled"
                  },
                  {
                    "id": 3,
                    "datetime": "2024-03-3T15:25:43.511Z",
                    "status": "pending"
                  }
                ]
        '204':
          description: The user has not made any orders
        '404':
          description: User not found
        default:
          description: Default error sample response
      tags:
        - Users
        - Orders
...